Reimplementation to fix races
This fixes a lot of races. The races didn't exist in the original C code
in 9pserve. They were introduced when the code was transliterated to
Go.

9pserve uses threads from libthreads, which run on one OS threads. You
can't simply translate libthread threads to goroutines, since goroutines
can be scheduled on multiple OS threads. To fix this, we redesign the
whole program instead of adding bunch of locks.

This new redesign simplified the data structures, removes reference
counting, and replacing queues with channels. It should be free of races
but may not be as faster as before (no message buffering using queues),
and may have introduces some bugs.

Fixes #2
Fixes #3